chore: GTK rendering/host split, resturcturing runtime projects
