+++ src/Uno.UI.Runtime.Skia.Gtk/Extensions/GtkApplicationViewExtension.cs
