+++ src/Uno.UI.Runtime.Skia.Gtk/Extensions/GtkCoreWindowExtension.cs
