+++ src/Uno.UI.Runtime.Skia.Gtk/Extensions/GtkDisplayInformationExtension.cs
