+++ src/Uno.UI.Runtime.Skia.Gtk/Extensions/System/LauncherExtension.cs
