+++ src/Uno.UI.Runtime.Skia.Gtk/Extensions/System/LauncherHelpers/MimeAppsList.cs
