+++ src/Uno.UI.Runtime.Skia.Gtk/Extensions/System/Profile/AnalyticsInfoExtension.cs
