+++ src/Uno.UI.Runtime.Skia.Gtk/Hosting/IGtkXamlRootHost.cs
@@ -1,0 +1,7 @@
+﻿#nullable enable
+
+namespace Uno.UI.Runtime.Skia.Gtk.Hosting;
+
+internal interface IGtkXamlRootHost
+{
+}
