--- src/Uno.UI.Runtime.Skia.Gtk/UI/Xaml/Controls/UnoGtkWindow.cs
@@ -1,8 +1,0 @@
-﻿namespace Uno.UI.Runtime.Skia.UI.Xaml.Controls;
-
-internal class UnoGtkWindow : Gtk.Window
-{
-	public UnoGtkWindow(string title) : base(title)
-	{
-	}
-}
