+++ src/Uno.UI.Runtime.Skia.Wpf/Rendering/IWpfRenderer.cs
@@ -9,6 +9,7 @@ internal interface IWpfRenderer : IDisposable
 	SKColor BackgroundColor { get; set; }
 
 	void Render(DrawingContext drawingContext);
+
 	bool Initialize();
 
 	SKSize CanvasSize { get; }
