+++ src/Uno.UI.Runtime.Skia.Wpf/UI/Controls/UnoWpfWindow.cs
