+++ src/Uno.UI/UI/Xaml/Hosting/IXamlRootHost.skia.cs
@@ -1,0 +1,20 @@
+﻿#nullable enable
+
+using Windows.UI.Xaml;
+
+namespace Uno.UI.Xaml.Hosting;
+
+internal interface IXamlRootHost
+{
+	bool IsIsland { get; }
+
+	UIElement? RootElement { get; }
+
+	XamlRoot? XamlRoot { get; }
+
+	void InvalidateRender();
+
+	void ReleasePointerCapture();
+
+	void SetPointerCapture();
+}
