+++ src/Uno.UI/Uno.UI.Skia.csproj
@@ -49,6 +49,10 @@
 	<Import Project="..\SourceGenerators\Uno.UI.Tasks\Content\Uno.UI.Tasks.targets" Condition="'$(SkipUnoResourceGeneration)' == '' " />
 
 	<ItemGroup>
+	  <UpToDateCheckInput Remove="UI\Xaml\Hosting\IXamlRootHost.skia.cs" />
+	</ItemGroup>
+
+	<ItemGroup>
 		<Service Include="{508349b6-6b84-4df5-91f0-309beebad82d}" />
 	</ItemGroup>
